Move struct to cpp
