+++ interconnect/interconnect.cpp
@@ -4,6 +4,30 @@
 #include <memory.h>
 #include <interconnect_internal.h>
 #include <iostream>
+#include <stree.h>
+
+
+typedef enum _bus_req_state
+{
+    NONE,
+    QUEUED,
+    TRANSFERING_CACHE,
+    TRANSFERING_MEMORY,
+    WAITING_CACHE,
+    WAITING_MEMORY
+} bus_req_state;
+
+typedef struct _bus_req {
+    bus_req_type brt;
+    bus_req_state currentState;
+    uint64_t addr;
+    int procNum;
+    uint8_t shared;
+    uint8_t data;
+    uint8_t dataAvail;
+    struct _bus_req* next;
+} bus_req;
+
 
 bus_req* pendingRequest = NULL;
 bus_req** queuedRequests;
@@ -26,6 +50,7 @@ static const char* req_state_map[] = {
 static const char* req_type_map[]
     = {[NO_REQ] = "None", [BUSRD] = "BusRd",   [BUSWR] = "BusRdX",
        [DATA] = "Data",   [SHARED] = "Shared", [MEMORY] = "Memory"};
+
 
 const int CACHE_DELAY = 10;
 const int CACHE_TRANSFER = 10;
